Added evaluations for cluster search
Plot of silhouette score for various numbers of clusters, visualization of clusters with t-SNE dimensionality reduction. Ideas for next steps are contained in evaluate script. The changes made to cluster search were just formatting things as functions so the scripts can work together and adjusting the number of clusters to match the best silhouette score